+++ sdk/plugins/youtube.cpp
@@ -1,3 +1,4 @@
+#include <QProcess>
 #include <openbr_plugin.h>
 
 #include "core/common.h"
@@ -16,16 +17,6 @@ class YouTubeFacesDBTransform : public UntrainableMetaTransform
     Q_PROPERTY(QString algorithm READ get_algorithm WRITE set_algorithm RESET reset_algorithm STORED false)
     BR_PROPERTY(QString, algorithm, "")
 
-    QSharedPointer<Transform> transform;
-    QSharedPointer<Distance> distance;
-
-    void init()
-    {
-        if (algorithm.isEmpty()) return;
-        transform = Transform::fromAlgorithm(algorithm);
-        distance = Distance::fromAlgorithm(algorithm);
-    }
-
     void project(const TemplateList &src, TemplateList &dst) const
     {
         Transform::project(src.mid(1) /* First template is the header in 'splits.txt' */, dst);
@@ -33,22 +24,19 @@ class YouTubeFacesDBTransform : public UntrainableMetaTransform
 
     void project(const Template &src, Template &dst) const
     {
+        static QMutex mutex;
         const QStringList words = src.file.name.split(", ");
-        dst.file.name = words[0] + "_" + words[1] + "_" + words[4] + ".mtx";
-
-        TemplateList queryTemplates = TemplateList::fromGallery(File(words[2]).resolved());
-        sort(queryTemplates);
-        queryTemplates >> *transform;
-
-        TemplateList targetTemplates = TemplateList::fromGallery(File(words[3]).resolved());
-        sort(targetTemplates);
-        targetTemplates >> *transform;
-
-        QScopedPointer<MatrixOutput> memoryOutput(MatrixOutput::make(targetTemplates.files(), queryTemplates.files()));
-        distance->compare(targetTemplates, queryTemplates, memoryOutput.data());
-
-        dst.clear();
-        dst.m() = memoryOutput.data()->data;
+        const QString matrix = "YTF-"+algorithm+"/"+words[0] + "_" + words[1] + "_" + words[4] + ".mtx";
+        const QStringList arguments = QStringList() << "-algorithm" << algorithm
+                                                    << "-parallelism" << QString::number(Globals->parallelism)
+                                                    << "-path" << Globals->path
+                                                    << "-compare" << File(words[2]).resolved() << File(words[3]).resolved() << matrix;
+        mutex.lock();
+        int result = QProcess::execute(QCoreApplication::applicationFilePath(), arguments);
+        mutex.unlock();
+        if (result != 0)
+            qWarning("Process for computing %s returned %d.", qPrintable(matrix), result);
+        dst = Template();
     }
 
     static void sort(TemplateList &templates)